+++ src/compute_uncertain_dim.jl
@@ -33,61 +33,100 @@ end
 
 
 
-function uncertainty_dimension_sample(xg, yg, basin)
 
+"""
+    uncertainty_exponent(bsn::basin_info, integ; sizes=0.)
+This function estimates the uncertainty exponent of the boundary. It is related to the uncertainty dimension.
+
+[C. Grebogi, S. W. McDonald, E. Ott, J. A. Yorke, Final state sensitivity: An obstruction to predictability, Physics Letters A, 99, 9, 1983]
+
+## Arguments
+* `bsn` : structure that holds the information of the basin.
+* `integ` : handle to the iterator of the dynamical system.
+
+## Keyword arguments
+* `sizes` array that specifies the scales at which the uncertainty exponent should be computed.
+
+"""
+function uncertainty_exponent(bsn::basin_info, integ; sizes=0.)
+    xg = bsn.xg; yg = bsn.yg;
     nx=length(xg)
     ny=length(yg)
     xi = xg[1]; xf = xg[end]; yi = yg[1];  yf = yg[end]
-    grid_res_x=length(xg)
-    grid_res_y=length(yg)
+    grid_res_x=xg[2]-xg[1]
+    grid_res_y=yg[2]-yg[1]
 
     N_u = [] # number of uncertain box
     N = [] # number of boxes
     ε = [] # resolution
 
-    min_ε = 5;
-    max_ε = grid_res_x/20
-    num_step=10
-    r,c= size(basin)
-    vals = unique(basin)
-    S=Int16(length(vals))
+
+    min_ε = 1e-6;
+    max_ε = grid_res_x*5;
+    num_step=15
+    r,c= size(bsn.basin)
+    vals = sum(isodd.(unique(bsn.basin)))
+    S=Int16(vals)
     if S < 2
         return 1,0,0
     end
 
     r_ε = range(min_ε,max_ε,length=num_step)
 
-    for (k,s_box) in enumerate(r_ε)
-        Nb=0; Nu=0;
+    for (k,eps) in enumerate(r_ε)
+        Nb=0; Nu=0; μ=0; σ²=0; M₂=0;
         completed = 0;
-        s_box = Int(round(s_box))
         # Find uncertain boxes
         while completed == 0
-            # Random box
-            x = rand(1:(r-s_box+1))
-            y = rand(1:(c-s_box+1))
-            x_coor=x:x+s_box-1
-            y_coor=y:y+s_box-1
-            box_values = [basin[k,m] for k in x_coor, m in y_coor]
-            Nb=Nb+1
-            tmp_Nu = Nu + (length(unique(box_values))>1)
-            if abs((Nu - tmp_Nu)/Nu) < 0.001 && Nu > r*c/s_box^2 && (length(unique(box_values))>1)
-                completed = 1
+            k1 = rand(1:length(xg))
+            k2 = rand(1:length(yg))
+
+            c1 = bsn.basin[k1,k2]
+            c2 = get_color_point!(bsn, integ, [xg[k1]+eps,yg[k2]])
+            c3 = get_color_point!(bsn, integ, [xg[k1]-eps,yg[k2]])
+
+            if length(unique([c1,Int(c2),Int(c3)]))>1
+                Nu = Nu + 1
             end
-            Nu = tmp_Nu
+            Nb += 1
+
+            # Welford's online mean and average estimation
+            M₂ = wel_var(M₂, μ, Nu/Nb, Nb)
+            μ = wel_mean(μ, Nu/Nb, Nb)
+            σ² = M₂/Nb
+
+            # Stopping criterion: variance of the estimator of the mean bellow 1e-3
+            if Nb > 100 && σ² < 1e-3
+                completed = 1
+                #@show Nu,Nb
+            end
+
         end
         push!(N_u,Nu)
         push!(N,Nb)
-        push!(ε,(xg[2]-xg[1])*s_box)
+        push!(ε,eps)
     end
+
     # uncertain function
     f_ε = N_u./N
 
     # get exponent
     @. model(x, p) = p[1]*x+p[2]
-    fit = curve_fit(model, vec(log.(ε)), vec(log.(f_ε)), [2., 2.])
+    fit = curve_fit(model, vec(log10.(ε)), vec(log10.(f_ε)), [2., 2.])
     D = coef(fit)
 
-    return D[1], ε, f_ε
+    return D[1]
+end
 
+
+
+function wel_var(M₂, μ, xₙ, n)
+
+    μ₂ = μ + (xₙ - μ)/n
+    M₂ = M₂ + (xₙ - μ)*(xₙ - μ₂)
+    return M₂
+end
+
+function wel_mean(μ, xₙ, n)
+    return μ + (xₙ - μ)/n
 end