+++ src/examples/tst_pendulum_all.jl
@@ -1,13 +1,18 @@
 using Revise
 using DifferentialEquations
+using DynamicalSystems
 using Basins
+using Plots
 using Printf
 
 # Equations of motion:
-function forced_pendulum!(du, u, p, t)
+function forced_pendulum(u, p, t)
+    @inbounds begin
     d = p[1]; F = p[2]; omega = p[3]
-    du[1] = u[2]
-    du[2] = -d*u[2] - sin(u[1])+ F*cos(omega*t)
+    du1 = u[2]
+    du2 = -d*u[2] - sin(u[1])+ F*cos(omega*t)
+    return SVector{2}(du1, du2)
+    end
 end
 
 # We have to define a callback to wrap the phase in [-π,π]
@@ -29,11 +34,11 @@ F = 1.66
 d=0.2
 p=[d, F, ω]
 #p=[0.15, 0.2, 0.1]
-df = ODEProblem(forced_pendulum!,rand(2),(0.0,20.0), p)
+df = ODEProblem(forced_pendulum,rand(2),(0.0,20.0), p)
 integ_df  = init(df, alg=AutoTsit5(Rosenbrock23()); reltol=1e-9, save_everystep=false, callback=cb)
 
-xres=100
-yres=100
+xres=200
+yres=200
 
 # range for forced pend
 xg = range(-pi,pi,length=xres)
@@ -43,10 +48,10 @@ yg = range(-2.,4.,length=yres)
 @time bsn = Basins.basin_map(xg, yg, integ_df; T=2*pi/ω)
 
 # Basin entropy
-@show Sb,Sbb = basin_entropy(bsn.basin; eps_x=20, eps_y=20)
+@show Sb,Sbb = basin_entropy(bsn; eps_x=20, eps_y=20)
 
 # Wada merge Haussdorff distances
-@time max_dist,min_dist = detect_wada_merge_method(xg, yg, bsn.basin)
+@time max_dist,min_dist = detect_wada_merge_method(xg, yg, bsn)
 epsilon = xg[2]-xg[1]
 @show dmax = max_dist/epsilon
 @show dmin = min_dist/epsilon
@@ -56,8 +61,12 @@ W = detect_wada_grid_method(integ_df, bsn; max_iter=8)
 @show W[:,end]
 
 # Uncertainty exponent for these parameter and grid
-bd = box_counting_dim(xg, yg, bsn.basin)
+bd = box_counting_dim(xg, yg, bsn)
 α = 2 - bd
+
+D = uncertainty_exponent(bsn, integ_df)
+@show 2-D
+
 
 println("---------------")
 println("---------------")
@@ -69,10 +78,9 @@ println("---------------")
 @printf("Boundary Basin Entropy: %.2f\n", Sbb)
 @printf("Uncertainty exponent: α= %.2f\n", α )
 @printf("Box counting dim: bd= %.2f\n", bd)
-
-@printf("Number of basins: %d\n", length(unique(bsn.basin)))
+@printf("Uncertainty dim estimator: d = %.2f\n", 2-D[1])
+@printf("Number of basins: %d\n", bsn.Na)
 @printf("Merge Method: Max fattening parameter: %.2f\n", dmax)
 @printf("Wada Grid Method: W_Na = %.2f\n ", W[end,end] )
-
 
 plot(xg, yg, bsn.basin', seriestype=:heatmap)